+++ html_form_builder/doc/changelog.rst
@@ -1,3 +1,7 @@
+v1.7.5
+======
+* Bug fix datetime field in backend
+
 v1.7.4
 ======
 * Record ID fix for external forms